Some crucial clk plumbing
clk-store-entry.pl takes a YAML-ish entry on STDIN and stores
a record for that entry in one of the timelines (doesn't yet
store the entry itself).

clk-in.pl generates a very simple entry (just a timestamp)
and sends it to clk-store-entry.pl for storage.

Eventually, I'd like to have clk-in.pl be something like this:

system("clk-make-entry @ARGV | clk-store-entry ./somewhere")
and
system("clk-report in")

make-entry converts the command line into a YAML-ish entry
store-entry stores it in the filesystem
the "in" report shows the last few entries