Favor newer timelines when storing entries
When looking for a timeline to hold the entry we're trying to add,
favor timelines whose oldest entry is more recent.  The main
motivation for this change is to make the behavior predictable and
therefore easily testable.

There may be performance implications.  If so, it should be easy
enough to address them when they arise.  Those implications should
only ever happen when someone is creating entries which are relatively
old.  That case is generally rare.