clk: provide only a default root path
For now, we still want clk setting the root path if
the user has not.  That allows me run basic commands
without prefixing everything with CLK_ROOT=...
However, the test suite sets the root path to something
for its own purposes.  If the root path is already
set, don't overwrite it.