Simple test for clk-in's current behavior
To make this work consistently, the tests must be able to
lie about the current time.  They do this by setting CLK_TIME
to the time in epoch seconds.  Presumably this will be useful
aside from testing.  For instance, if someone wants to write
an import script which drives 'clk-in', it could just set
CLK_TIME to the necessary time before running the script.