Rudimentary search between two times
We can now search for entries which happened between two time
periods.  This is the most rudimentary search between two times
that I can think of.  The next step is to add more tests to
between.t to make sure that the code correctly handles multiple
timelines.  After that, parsing should be added so that the bounds
between which searches are performed can be specified in any "instant"
format.  After that, we can add a --period argument which converts
an English-language period into two epoch seconds and then invokes
the same search by time.

Signed-off-by: Michael Hendricks <michael@ndrix.org>