Don't use FindBin
FindBin is much too heavyweight for what we need.  By removing it, we
reduced the number of stat() system calls by 50% down to around 200
when running `clk list`.  It should be easy enough to just replace the
hard-coded constant at build time since we'll be installing the clk
program in a specific location.

Signed-off-by: Michael Hendricks <michael@ndrix.org>