Include microseconds in the time stamp
It's easy enough to include microseconds that it might as well be
included from the outset.

The idea of this rough program is to store entries in a giant time
line (one time line per "identity").  Each entry records a message and
some tags along with it.  The output of clk-in.pl is suitable for
storing in a file with one record per line.  A nice benefit of this
arrangement is that "cat time-line-A time-line-B | sort > time-line"
can merge two time lines.

It will probably make sense to add whitespace padding at the end of
each line so that the time line file can be treated as a list of fixed
width records.

Signed-off-by: Michael Hendricks <michael@ndrix.org>