Quick prototype with a storage interface
This is just a quick hack showing roughly how a standard storage
interface could be used for implementing a rudimentary clk version.
If this naïve SQLite version ever needed to be replaced with something
more robust, it could be easily replace with any other implementation
of the same interface.

Ideally, I'd like to make the interface a Haskell type class and
implement various types in that class to provide the implementation.
It'd be a good exercise in learning Haskell and would also make it
easier to distribute by compiling binaries for various platforms.