New 'list' command
This command lists the last three entries that were created.  This
implementation is so ridiculously inefficient that it never returns on
an entries log with 1,000,000 entries.  Presumably it reads the entire
file into memory, creates the Entry values, finds the last three and
prints them.

This should do something smarter like seeking toward the end of the file
and extracting the last three lines that way.