a rough working sketch on where to go next
The idea is that a File is a list of Hunks is a list of lines.  deepTake
traverses this nested list of lists to take out the right number of
lists.  squish takes a list of lines and combines adjacent partial
lines.  If I want 5 lines, the naive approach is

    squish $ deepTake 5 file

however, with partial lines some of those 5 taken lines may be combined
together.

It seems like the best way forward is to modify deepTake so that it
knows how to combine adjacent partial lines while its taking.  It should
be able to do it by traversing the list two at a time (or something
similar).