During testing, using my business identity
It doesn't really matter, but since I use clk for work (predominantly),
I should probably use that email address.

The next step is to write a 'dump' command which does the equivalent of
sending events.log to stdout by retrieving events from the storage
engine and printing them in the standard, tab-separated form.

That will make list and dump commands do almost the exact same thing.
The list command should then be cleaned up a bit so that it looks more
like:

    0d00eb 09/07 13:59                 out
    72cee9 09/07 14:13 ndrix,research  reading blogs
    841ca7 09/07 14:17 ndrix,20        clk

The important parts are the entry ID, shortened time format and column
alignments.