Rudimentary pre-in hook handling
'clk in' runs the pre-in hook passing the Event on stdin.  The hook must
return on stdout the exact Event that should be added to storage.  It
works if everything is done right.  If not, the error messages are hard
to understand.

With a quickly hacked together pre-in hook like

    #!/usr/bin/perl
    # do nothing as a pre-in hook
    chomp( my $entry = <> );
    my ($who, $when, $subject, $tags) = split /\t/, $entry;
    my @tags = split /,/, $tags;

    if ( $subject eq 'gsg email' ) {
        $subject = 'email';
        @tags = 'gsg';
    }

    print join "\t", $who, $when, $subject, join(',',@tags);
    print "\n";
    exit 0;

I can run

    ./clk in gsg email

and get

    michael@ndrix.com       2009-09-11T22:44:58.136847Z     email   gsg

in the entries log.  Hooray!