A rough dump command
This command only understands the period 'today' but it shows the
correct output for except for the last event.  Running

 ./clk dump -p today

gives results like

 ...
 michael@ndrix.com       2009-09-18T20:29:22.270534Z     testing clk 20
 clk: Can't create an Event from: ["michael@"]

The Event parser seems to be confused by the very last entry.  I'm not
entirely sure what's causing the problem.  Maybe I should try adding a
blank line to the end of the file or something and see if that changes
the output.