Support tags when clocking in
Any words on the command line of the form ".foo" are treated as tags and
recorded as such in the timeline.  I would have liked to use "#foo" (ala
Twitter), but bash treats any text on the commandline following # as a
comment ignores it.  That makes for very ineffective tagging ;-)  The
leading period seemed reasonable since it's easy to type and almost
never (ever?) used under normal circumstances.