Very rough start on clk-ls
The goal of this script is to list all the entries for the most recent
time period (currently defined as a month).  It will indicate the
duration of each entry.  The duration is the key part of the output.
Anything else is frosting on the cake.

The duration of an entry is the time between this entry and the one
following it.