Parse entries from an entry line
This code isn't actually used yet, but it works as expected in ghci.  It
takes a single entry line (as created by `in`) and converts it into an
Entry.  This is the first time an Entry data type has been used.  I
tried hacking together some code without this and it became conceptually
too complicated to work with.  Hopefully encapsulating entries behind
this data type will help.

One note: read does not expect a trailing newline on the entry line.
The trailing newline is a delimiter used when storing the entry to disk
and will be trimmed off during actual usage.