Make Entry a custom Show instance
The instance from "deriving" was there only for debugging.  This
implementation does the same thing as clk-in.  The duplication between
those two places should be factored out soon.

At this point, I've made the decision that show and read deal with the
filesystem representation of an Entry.  That seems reasonable since the
two are inverses of each other.  Displaying an Entry for user
consumption only needs to go one way.