An Entry might have a duration
If an Entry value is created in isolation, it won't have a sensible
duration.  A duration is only reasonably when an entry occurs alongside
a second entry.  In that case, the duration could be calculated.

For now, we don't ever calculate the duration.  We just leave a place
for it in the Entry constructor.