Include duration in `ls` output
For each entry, calculate how long that entry lasted (based on the
entry's context) and display that duration in a user-friendly,
high-level summary sort of way.

This code has one *significant* bug.  It doesn't display a line for the
final entry.  That needs to be fixed.  It's important to be able to see
the most recent entry.