Sum duration by client
Calculate the sum of entry durations grouped by client.  The output
format is whatever Haskell chooses to display for Data.Map  I'll change
that shortly.

Ideally this report would be configurable but since grouping by client
is the one that I run most frequently, I'm hard coding it for now.