Factor out code for display hour-long durations
Durations as seconds and fractions thereof (the default show format for
NominalDiffTime) isn't particularly useful in the context of a time
clock application.  The new function showDurationAsHours displays the
duration as a number of hours with precision out to 2 decimal points
(precision of about half a minute).  Numbers of this magnitude are more
useful in this context.