Restrict report output to today
Instead of summarizing time worked for all entries, do it for those
entries that started today (local time).  This isn't always what I want,
but it's what I want for the next 6 days or so.

In one of the other branches is some code for converting English
language phrases ("today", "this week", etc) into a pair of UTCTime
values.  Eventually I'll use that module to determine the end points and
report on entries that fall between them.