Correctly find the most recent month file
We previously relied upon there only being one file in the timeline
directory or the OS returning the files in reverse chronological order.
Now we sort the month files ourself to guarantee that the file we choose
is the most recent one.

This makes clk-ls and clk-report actually work after importing historic
timeline data from the old clk program.