Import Data.Period from the haskell branch
This abstraction of a period of down bounded on both ends is quite
handy.  The parsing routines for converting English phrases into UTCTime
bounds is perhaps the most useful part of it.

I added the function 'within' after importing the code and used record
syntax to define Period to give us easy accessors for the components.