Allow non-now references for an enclosing period
When calculating the bounds of a Period value, we have previously used
the current time as the point of reference.  Namely, we find the period
of the desired size which encloses the current time.  Unfortunately,
that won't work for phrases like "yesterday" or "last week" which don't
enclose the current time.

We now define a period in terms of a desired duration (day, week, month,
etc) and a duration into the future or past.