Start of System.AbsolutePath
This type is a restriction of FilePath which guarantees that the path is
absolute.  Perhaps there are better ways of handling this, but I find it
conceptually valuable to know that a path I have is guaranteed to be
absolute and I can use it without regard to the current directory or a
specific base directory.

A couple functions implemented by FilePath are also implemented.  They
just pass through to the FilePath function.  It'd sure be handy if
Haskell had subtypes like Moose.  Then I could do

    subtype 'AbsolutePath'
        => as 'FilePath'
        => where { isAbsolute $_ }
        => message { "The path $_ is not absolute" }