Start on App.Clk.MonthFile
This type represents files on disk which contain clk entries for an
entire month.  It's convenient to bundle together the absolute path to
the file (so it can be read or written to) and the time period covered
by the file (so it can be discarded quickly during searches).

The main interface to this type is maybeMonthFile which takes an
absolute path which might represent a MonthFile and possibly constructs
a value of that type if it is.