Add Show, Eq and Ord instances for MonthFile
The Show instance is handy for debugging in ghci.  The Eq instance is
required for the Ord instance.  The Ord instance will be useful when
trying to determine what is the most recent month file in the timeline.

I switched MonthFile to record syntax because compare is clearer when I
have the period accessor.