DCG describing marks recorded in a text file
Each time a user clocks in, we create a mark on the timeline.  The mark
contains a timestamp and a note (space separated list of words).  These
DCG predicates describe this text format.  They can both parse and
generate mark lines based on mark values.

There's also supported for an antiquated mark line format in which a
username and tags field existed.