Factor out file_time/2
This predicate describes (half of) the relationship between a timeline
file and a julian datetime.  Because a datetime can describe multiple
moments in history, it's possible that multiple timeline files could
cover it.

Of course, just because the datetime covers some moments in time
doesn't mean that we actually have a timeline for that moment.  That
explains the exists_file/1 call.